Rename server files to snake case
