+++ src/RobotServer/single_image_capture.py
